pyproject.toml
[project]
name = "nexus-trade-engine"
version = "0.1.0"
description = "Algorithmic trading backtesting and execution engine"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "polars>=1.16.0",
    "taskiq>=0.11.0",
    "taskiq-fastapi>=0.3.0",
    "taskiq-redis>=1.0.0",
    "valkey>=6.0.0",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "sentry-sdk[fastapi]>=2.18.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
    "orjson>=3.10.0",
    "numpy>=2.4.4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "python-ldap>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.5.0",
    "hypothesis>=6.115.0",
    "ruff>=0.8.0",
    "basedpyright>=1.22.0",
    "httpx>=0.28.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["engine"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "S", "B", "A", "C4",
    "DTZ", "T20", "ICN", "PIE", "PT", "RSE", "RET", "SLF",
    "SIM", "TID", "TCH", "ARG", "ERA", "PGH", "PL", "TRY",
    "FLY", "PERF", "RUF",
]
ignore = ["S101", "TRY003", "PLR0913"]

[tool.ruff.lint.isort]
known-first-party = ["engine"]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportMissingTypeStubs = false
reportUnknownMemberType = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra --strict-markers --cov=engine --cov-report=term-missing --cov-report=html --cov-fail-under=80"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
]
